Blue label
In draw(), the bubbles rise and get bigger over time. It should be reset to the bottom when it reaches the top of the bottle.

The blue label part of the bubble should be invisible because the label is opaque. After the bubble is drawn in the draw() function, redraw the label.